[deployment] Fix race condition which makes deployments instantaneous
If the latency between the server running Terraform and the K8s masters
is low enough a race condition will prevent Terraform to wait for the
deployment rollout to succeed regardless of the value of the
wait_for_rollout variable.
This is apparently caused by Terraform getting the id of the
specReplicas in the waitForDeploymentReplicasFunc function before the
new value is propagated to the Deployment object on the K8s masters.